+++ lib/bumblebee/text/blip_text.ex
@@ -154,7 +154,7 @@ defmodule Bumblebee.Text.BlipText do
   @impl true
   def input_template(_spec) do
     %{
-      "input_ids" => Nx.template({1, 1}, :s64)
+      "input_ids" => Nx.template({1, 1}, :u32)
     }
   end
