+++ lib/bumblebee/text/mbart.ex
@@ -229,7 +229,7 @@ defmodule Bumblebee.Text.Mbart do
   @impl true
   def input_template(_spec) do
     %{
-      "input_ids" => Nx.template({1, 1}, :s64)
+      "input_ids" => Nx.template({1, 1}, :u32)
     }
   end
